Implement repetitions code

+++ app/src/main/java/htw/gma_sose22/metronomekit/metronome/Metronome.kt
@@ -4,7 +4,7 @@ import htw.gma_sose22.metronomekit.beat.NextToneProvider
 import htw.gma_sose22.metronomekit.beat.Tone
 
 class Metronome(
-    override var bpm: Int = DEFAULT_SPEED.toInt(),
+    override var bpm: Int = DEFAULT_SPEED,
     override var beatSound: ByteArray,
     override var offbeatSound: ByteArray,
     override val metronomeAudio: MetronomeAudioInterface,
@@ -13,7 +13,7 @@ class Metronome(
 
     companion object {
         const val DEFAULT_SAMPLE_RATE = 44100
-        const val DEFAULT_SPEED = 120u
+        const val DEFAULT_SPEED = 120
     }
 
     override var isPlaying = false
